devices: Tests to list devices
